Refactor,fix typo
