+++ helloworld.py
@@ -1,14 +1,14 @@
 import slack
 import slackhelper
-import slackbot
+import polypbot
 import schedule
 import time
 
 
 def start_daily_stand_up():
     _slack = slackhelper.SlackHelper()
-    _slack.send_notification_to_users(message=slackbot.PolypBot.GREETINGS,
-                                      blocks=slackbot.PolypBot.GREETINGS_BLOCK)
+    _slack.send_notification_to_users(message=polypbot.Constants.GREETINGS,
+                                      blocks=polypbot.Constants.GREETINGS_BLOCK)
     # schedule.cancel_job(start_daily_stand_up)
     _slack.start_rtm()
