+++ slackhelper.py
@@ -1,15 +1,15 @@
 import slack
-import daily
+import polypbot
 import constants
 
 
 class SlackHelper:
 
     def __init__(self):
-        self.slack_token = 'xoxb-761143478773-760710426068-TBDVs3VQgt3PUMmDtZmYvjfj'  # bot_token
+        self.slack_token = 'xoxb-761143478773-760710426068-6C0NhW3simtboAzSRHEFPnaH'  # bot_token
         self.client = slack.WebClient(token=self.slack_token)
         self.rtm_client = slack.RTMClient(token=self.slack_token)
-        self.users = {}
+        self.check_ins = {}
 
     def start_rtm(self):
         self.rtm_client.start()
@@ -36,20 +36,21 @@ class SlackHelper:
         users_list = self.client.users_list().get('members', None)
         if users_list:
             for u in users_list:
-                print(u)
                 try:
                     user_id = u['id']
                     if user_id and user_id not in constants.CHECK_INS and not u['deleted'] and not u['is_bot']:
-                        message_ = message.format(user=user_id, dept='IT DEPT')
+                        message_ = message.format(user=user_id, dept='IT DEPT',
+                                                  first_standup_question=polypbot.Constants.STAND_UP_QUESTIONS[1])
                         blocks[0]['text']['text'] = message_
                         # print('{blocks}'.format(blocks=blocks))
                         if self.post_message_with_block_template(uid=user_id,
                                                                  blocks=blocks):
-                            constants.CHECK_INS[user_id] = daily.Standup(account_id=user_id,
-                                                                         account_name=u.get('profile').get('real_name',
-                                                                                                           'polypbot'),
-                                                                         photo_url=u.get('profile').get('image_48',
-                                                                                                        ''))
+                            constants.CHECK_INS[user_id] = polypbot.Standup(account_id=user_id,
+                                                                            account_name=u.get('profile').get(
+                                                                                'real_name',
+                                                                                'polypbot'),
+                                                                            photo_url=u.get('profile').get('image_48',
+                                                                                                           ''))
                 except KeyError:
                     pass
 
@@ -64,10 +65,6 @@ class SlackHelper:
             user_check_in = constants.CHECK_INS[user_id]
             user_check_in.add_user_answer(data.get('text'))
             if user_check_in.user_answered_all_questions and not user_check_in.is_done:
-                print('{user_id}, is_done={done},answered all = {ans} \n {res}'.format(user_id=user_id,
-                                                                                       done=user_check_in.is_done,
-                                                                                       ans=user_check_in.user_answered_all_questions,
-                                                                                       res=user_check_in.result))
                 # self.post_message(channel='#random', message='tapos na si <@{user}>!'.format(user=user_id))
                 web_client.chat_postMessage(
                     channel='#random',
@@ -78,9 +75,9 @@ class SlackHelper:
                 )
                 user_check_in.is_done = True
             elif not user_check_in.user_answered_all_questions:
-                # self.post_message(channel='#???????', message=user.get_next_stand_up_question)
+                # self.post_message(channel=user_id, message=user.get_next_stand_up_question)
                 web_client.chat_postMessage(
                     channel=user_id,
-                    text=user_check_in.get_next_stand_up_question,
+                    text='*{}*'.format(user_check_in.get_next_stand_up_question),
                     as_user=True,
                 )
